Refactor Slurm code into cw.slurm module.
Now it is possible to programmatically start/stop a worker cluster from
a script (see square.py example).

Fixes #6.
